+++ battlecode22-scaffold/src/ATopTenBot/Util.java
@@ -122,7 +122,7 @@ public class Util extends Globals {
 		RobotInfo closestUnit = null;
 		int minDistSq = Integer.MAX_VALUE;
         int distSq = 0;
-        MapLocation lCR = currentLocation;
+        MapLocation lCR = rc.getLocation();
 		for (int i = units.length; i --> 0; ) {
 			distSq = lCR.distanceSquaredTo(units[i].location);
 			if (distSq < minDistSq) {