+++ battlecode22-scaffold/CHANGELOG.md
@@ -1,14 +1,24 @@
-# PageOneBot
+# CHANGELOG
+
+## PageOneBot
+
 * Implement archon fleeing, may tweak value and other stuff regulating when it flees
 * Soldiers will now update comms with enemy archon location
 
-# FoolsGoldBot
+## FoolsGoldBot
+
 * Reintroduced update_enemy_archon location for soldiers
 * Some tweaking to the retreat function calculation
 * More accurate enemy locations and skip our own archon locations for small maps
 
-# FoolsGoldNewBot:
+## FoolsGoldNewBot:
+
 * Added reset miningLocation and inPlaceToMine when fleeing from enemy robot.
 
-# GrindstoneBot:
+## GrindstoneBot:
+
 * Added offensive Archon repositioning code.
+
+## TopTenBot
+
+* Change all currentLocation mentions to rc.getLocation() as only 1 bytecode cost